+++ shared/package.json
@@ -3,5 +3,8 @@
   "private": true,
   "dependencies": {
     "@aws-sdk/client-s3": "3.1015.0"
+  },
+  "scripts": {
+    "test": "bun test"
   }
 }
